+++ esphome/components/microphone/automation.h
@@ -16,10 +16,10 @@ template<typename... Ts> class StopCaptureAction : public Action<Ts...>, public 
   void play(Ts... x) override { this->parent_->stop(); }
 };
 
-class DataTrigger : public Trigger<const std::vector<uint8_t> &> {
+class DataTrigger : public Trigger<const std::vector<int16_t> &> {
  public:
   explicit DataTrigger(Microphone *mic) {
-    mic->add_data_callback([this](const std::vector<uint8_t> &data) { this->trigger(data); });
+    mic->add_data_callback([this](const std::vector<int16_t> &data) { this->trigger(data); });
   }
 };
